packages/shared/pyproject.toml
[project]
name = "shared"
version = "0.1.0"
description = "Shared utilities for LIBI ecosystem"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.23",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.25.0",
    "tenacity>=8.2.0",
    "structlog>=23.2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
llm = [
    "anthropic>=0.8.0",
    "openai>=1.6.0",
    "google-generativeai>=0.3.0",
]
cloud-aws = [
    "boto3>=1.34.0",
]
cloud-gcp = [
    "google-cloud-storage>=2.14.0",
    "google-cloud-secret-manager>=2.17.0",
]
cloud-azure = [
    "azure-storage-blob>=12.19.0",
    "azure-keyvault-secrets>=4.7.0",
    "azure-identity>=1.15.0",
]
tts = [
    "google-cloud-texttospeech>=2.14.0",
    "elevenlabs>=0.2.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]
all = [
    "shared[llm,cloud-aws,cloud-gcp,cloud-azure,tts,dev]",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["shared*"]
